requirement.txt
tensorflow
streamlit
numpy
pillow
matplotlib
scikit-learn
